Various fixes.

+++ TOMB4/game/trng/trng_flipeffect.cpp
@@ -755,9 +755,9 @@ bool static_move_static_with_data_in_x_parameter_list(unsigned char move_param_i
 	switch (direction) {
 		case DIR_NORTH: {
 			if (!NGGetStaticHorizontalMovementRemainingUnits(script_static)) {
-				short current_angle = 0xC000;
+				short current_angle = (short)0xC000;
 				if (move_item->direction & DIR_INVERT_DIRECTION) {
-					current_angle += 0x8000;
+					current_angle += (short)0x8000;
 				}
 				NGSetStaticHorizontalMovementAngle(script_static, (short)current_angle);
 				NGSetStaticHorizontalMovementRemainingUnits(script_static, move_item->distance);
